requirements.txt
python-dotenv
openai
neo4j
langchain
langchain_community
psycopg2
PyPDF2